perform nft query only once when fetching transactions with operations
