+++ src/endpoints/transactions/transaction.service.ts
@@ -424,6 +424,8 @@ export class TransactionService {
       detailedTransactions.push(transactionDetailed);
     }
 
+    await this.transactionGetService.applyNftNameOnTransactionOperations(detailedTransactions);
+
     return detailedTransactions;
   }
 
@@ -488,6 +490,8 @@ export class TransactionService {
         operations.push(undefined);
       }
     }
+
+    await this.transactionGetService.applyNftNameOnTransactionOperations(transactions);
 
     return operations;
   }
